Fix VS2015 issues

+1: Paul Novotny, Ryan Beasley, Gughan Sathyaseelan

Merged-on: https://assembla.com/code/OpenSurgSim/git/merge_requests/2361113

+++ SurgSim/Graphics/TextRepresentation.cpp
@@ -50,8 +50,8 @@ TextRepresentation::TextRepresentation(const std::string name) : Representation(
 	SURGSIM_ADD_SERIALIZABLE_PROPERTY(SurgSim::Graphics::TextRepresentation, Math::Vector4d, BackgroundColor,
 									  getBackgroundColor, setBackgroundColor);
 
-	SURGSIM_ADD_SERIALIZABLE_PROPERTY(SurgSim::Graphics::TextRepresentation, float, BackgroundMargin,
-									  getBackgroundMargin, setBackgroundMargin);
+	SURGSIM_ADD_SERIALIZABLE_PROPERTY(SurgSim::Graphics::TextRepresentation, double, BackgroundMargin,
+ 									  getBackgroundMargin, setBackgroundMargin);
 
 }
